+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.0.42
 
+- Add vertical/horizontal scrollbars to Meta-Sample
+- Keep only vertical scrollbar in Hello, Text!
 - Fix issue with including bwidgets and iwidgets in Windows
 - Fix Hello, Menu Bar! sample in Windows (by disabling Windows-special System menu, which does not seem to be working in Tk)